Move get_value_for_tag to more general utils/misc
